pyproject.toml
[tool.poetry]
name = "terminalbackend"
version = "0.1.0"
description = ""
authors = ["Sergey Nesterov <sergioreklamplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
redis = "^5.2.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytoniq = "^0.1.40"
pytoniq-core = "^0.1.39"
asyncio = "^3.4.3"
pytvm = "^0.0.14"
base58 = "^2.1.1"
psycopg2 = {version = "^2.9.10", platform = "win32"}
psycopg2-binary = {version = "^2.9.10", platform = "linux"}
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
